requirements.txt
langchain
streamlit
faiss-cpu
sentence-transformers
python-dotenv
